Implement various changes to enhance Python 2.7 compatibility
This commit is the result of squashing several intermediate commits.
The individual commit messages were as follows:

Commit message 1
----------------

Roll back changes from commit 4910b785.

The commit referenced above removed uses of `socket.error`,
preferring instead to use the Python 3 builtin error types.
This causes various subtle issues with Python 2.7.

In this "revert", I have opted to have the compatibility
`BlockingIOError` be defined as a subclass of IOError, rather than
simply a name alias, because at least within this file, at least,
all uses where we need to check for it. Since `socket.error` is a
subclass of IOError, making `BlockingIOError` a name alias would make
the code more confusing, since it may not be obvious to the reader that
this is the case.

Commit message 2
----------------

Bump __version__ to reflect Digi patches have been applied

Commit message 3
----------------

Add _on_pre_connect default value of None

Otherwise a client using `.connect_async` will error out.

Commit message 4
----------------

Catch and handle socket.error on reconnect calls

Commit message 5
----------------

Use select.poll instead of select.select

select.select only works if the fileno values of the given objects
are <1024. In an application that uses or creates hundreds of file
descriptors, one can very easily end up in a situation where the socket
file descriptor is beyond that.

A typical Linux system will have the open file limit for the user set
to 1024, so this generally doesn't happen. But with this change in
place, Paho will function correctly if it does.